Fix signed/unsigned comparison warning
git-svn-id: svn://scribus.net/trunk/Scribus@23248 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/text/index.cpp
@@ -14,7 +14,7 @@
 // find run with runStart <= pos < runEnd
 uint RunIndex::search(int pos) const
 {
-	std::vector<uint>::const_iterator it = std::upper_bound(runEnds.begin(), runEnds.end(), pos);
+	std::vector<uint>::const_iterator it = std::upper_bound(runEnds.begin(), runEnds.end(), unsigned(pos));
 	return it - runEnds.begin();
 }
